+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/notifications/Notification.kt
@@ -42,8 +42,9 @@ data class Notification(
         fun emptyDiff() = Notification(DEFAULT_TITLE, message = message("notifications.empty-diff"))
         fun promptTooLarge() = Notification(DEFAULT_TITLE, message = message("notifications.prompt-too-large"))
         fun unsuccessfulRequest(message: String) = Notification(message = message("notifications.unsuccessful-request", message))
-        fun noCommitMessage(): Notification = Notification(message = message("notifications.no-commit-message"))
-        fun unableToSaveToken(): Notification = Notification(message = message("notifications.unable-to-save-token"))
+        fun noCommitMessage() = Notification(message = message("notifications.no-commit-message"))
+        fun unableToSaveToken() = Notification(message = message("notifications.unable-to-save-token"))
+        fun noCommonBranch() = Notification(message = message("notifications.no-common-branch"))
 
     }